[FIX] Specify width of columns in Trial Balance report
After making the space character the thousands separator,
numbers are no longer one "word". This allowed the system
to widen column 2 to fit more on one line, at the expense
of the width of the last 6 which held the numbers. As a
result, the numbers got split over multiple lines.